+++ zerver/tests/test_users.py
@@ -774,7 +774,7 @@ class QueryCountTest(ZulipTestCase):
                         acting_user=None,
                     )
 
-        self.assert_length(queries, 70)
+        self.assert_length(queries, 71)
         self.assert_length(cache_tries, 22)
         self.assert_length(events, 7)